SPRINT#2 (V0.3.3 & V0.3.4) - 0.3.3 - Query good for settings but also added region identifiers to the URL to know what level to load in the campain version of the simulator (V2.2.3).
0.3.4 - Done backend of the level checking and data. Bugs fixed by also making and debugging the corrosponding tests (heurestic testing included) (V2.2.4).

Level objective appears to be too large for the display box element so the display box will instead display the input/output case example while the level objective will be underneath it and the status to have more horizontal space. This is part of preparation of linking the level checking backend with the front end.