For better user testing, configured github repository to act as a website.
To do this I need to make index.html in the root folder which simply directs to a different HTML file. Used code from https://www.tutorialspoint.com/How-to-use-Meta-Tag-to-redirect-an-HTML-page as template. If the assumed redirecting did not work then user can still manually redirect by listening to the only displayed text and clicking on the hyperlink.